bevy_sprite
Function prepare_materials_2d

pub fn prepare_materials_2d<M: Material2d>(
    prepare_next_frame: Local<'_, PrepareNextFrameMaterials<M>>,
    extracted_assets: ResMut<'_, ExtractedMaterials2d<M>>,
    render_materials: ResMut<'_, RenderMaterials2d<M>>,
    render_device: Res<'_, RenderDevice>,
    images: Res<'_, RenderAssets<Image>>,
    fallback_image: Res<'_, FallbackImage>,
    pipeline: Res<'_, Material2dPipeline<M>>,
)

This system prepares all assets of the corresponding Material2d type which where extracted this frame for the GPU.